backend/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Pillow==8.3.2
psycopg2==2.9.3
pycodestyle==2.7.0
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==6.2.0
